feat(example) create auth provider example
This change adds an example showing a way to install an authentication
provider using the dynamic plugins feature of RHDH.  The example
includes two wrappers that make use of an existing set of plugins and a
guide that walks through building the example and setting it up.  The
guide also includes instructions on how to set up a simple LDAP server
to see this all running in a local environment.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>